* Added public ipfs route under basic auth * Changed direction field side in contract to int16

+++ contracts/VideoRequester.sol
@@ -16,13 +16,13 @@ contract VideoRequester is ChainlinkClient, ConfirmedOwner {
         uint32 long;
         uint32 start;
         uint32 stop;
-        uint8 direction;
+        uint16 direction;
         uint reward;
         string url;
         address requester;
     }
 
-    event VideoRequested(address requester, uint32 lat, uint32 long, uint32 start, uint32 end, uint8 direction, uint reward);
+    event VideoRequested(address requester, uint32 lat, uint32 long, uint32 start, uint32 end, uint16 direction, uint reward);
     event VideoReceived(string url);
 
     mapping (string => VideoRequest) requests;
@@ -35,7 +35,7 @@ contract VideoRequester is ChainlinkClient, ConfirmedOwner {
         fee = (1 * LINK_DIVISIBILITY) / 10; // 0,1 * 10**18 (Varies by network and job)
     }
  
-    function submitRequest (string memory id, uint32 lat, uint32 long, uint32 start, uint32 end, uint8 direction) payable external {
+    function submitRequest (string memory id, uint32 lat, uint32 long, uint32 start, uint32 end, uint16 direction) payable external {
         emit VideoRequested(msg.sender, lat, long, start, end, direction, msg.value);
     
         requests[id] = VideoRequest(id, lat, long, start, end, direction, msg.value, "", msg.sender);
